requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.6
Flask-Mail==0.9.1
pywebview
pyinstaller 